feat(sever): split logger creation to a simple function to allow non-dependecy-injection usage
